implemented own map renderer, closed #31 .

+++ engine/src/main/java/de/leeksanddragons/engine/renderer/BasePage.java
@@ -25,12 +25,12 @@ public abstract class BasePage implements IPage {
     }
 
     @Override
-    public float getWidth() {
+    public int getWidth() {
         return this.width;
     }
 
     @Override
-    public float getHeight() {
+    public int getHeight() {
         return this.height;
     }
